third_party/blink/web_tests/fast/scroll-snap/animate-fling-to-snap-points-1.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/gesture-util.js"></script>
<style>
div {
  position: absolute;
}
#scroller {
  width: 500px;
  height: 500px;
  overflow: scroll;
  scroll-snap-type: both mandatory;
  border: solid black 5px;
}
#space {
  width: 2000px;
  height: 2000px;
}
.target {
  width: 200px;
  height: 200px;
  scroll-snap-align: start;
  background-color: blue;
}
</style>

<div id="scroller">
  <div id="space"></div>
  <div class="target" style="left: 0px; top: 0px;"></div>
  <div class="target" style="left: 80px; top: 80px;"></div>
  <div class="target" style="left: 200px; top: 200px;"></div>
</div>

<script>
if (window.internals) {
  internals.runtimeFlags.implicitRootScrollerEnabled = true;
  // Jump directly to the snap position.
  internals.settings.setScrollAnimatorEnabled(false);
}

var body = document.body;
var scroller = document.getElementById("scroller");
var space = document.getElementById("space");

function scrollLeft() {
  return scroller.scrollLeft;
}

function scrollTop() {
  return scroller.scrollTop;
}

// MacOS devices do not have touch screens. Tests are irrelevant.
const macPlatform = navigator.platform.indexOf('Mac') == 0;

function reset() {
  return waitForScrollReset(scroller);
}

async function scrollUpdate(scroll_trigger) {
  const scrollPromise = waitForScrollendEvent(scroller);
  await scroll_trigger();
  return scrollPromise;
}

promise_test (async () => {
  if (macPlatform)
    return;

  await reset();
  await scrollUpdate(async () => {
    // 'downright' is a positive scroll in both the x and y direciton.
    await smoothScroll(100, 200, 200, GestureSourceType.TOUCH_INPUT,
                 'downright', SPEED_INSTANT)
  });
  assert_equals(scrollLeft(), 80);
  assert_equals(scrollTop(), 80);
}, "Without fling enabled, the scroll ends at the closest snap point to the " +
   "scroll destination.");

promise_test (async () => {
  await reset();
  await scrollUpdate(async () => {
    // Swipe up and to the left to scroll down and to the right.
    await swipe(100, 200, 200, 'upleft', 900);
  });
  assert_equals(scrollLeft(), 200);
  assert_equals(scrollTop(), 200);
}, "Fling scroll ends at the closest snap point to the fling destination.");

promise_test (async () => {
  await reset();
  await scrollUpdate(async () => {
    // Swipe left to scroll right.
    await swipe(100, 200, 200, 'left', 900);
  });
  assert_equals(scrollLeft(), 200);
  assert_equals(scrollTop(), 0);
}, "Fling scroll in one direction ends at closest snap point to the fling " +
   "destination.");
</script>
